+++ src/test/java/cart/service/UserServiceTest.java
@@ -1,18 +1,12 @@
 package cart.service;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.mockito.BDDMockito.given;
 
 import cart.domain.user.User;
-import cart.exception.InvalidPasswordException;
-import cart.exception.UserNotFoundException;
 import cart.repository.user.UserRepository;
 import java.util.List;
-import java.util.Optional;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -40,52 +34,5 @@ class UserServiceTest {
         List<User> result = userService.findAllUser();
         //then
         Assertions.assertThat(result).contains(user, otherUser);
-    }
-
-    @Nested
-    @DisplayName("유저 검증 기능")
-    class AuthorizeUser {
-
-        @Test
-        @DisplayName("유저를 검증할 수 있다.")
-        void authorize_user_success() {
-            //given
-            String email = "a@a.com";
-            String password = "password";
-            given(userRepository.findByEmail(email)).willReturn(Optional.of(new User(email, password)));
-
-            //when
-            //then
-            assertDoesNotThrow(() -> userService.authorizeUser(email, password));
-        }
-
-        @Test
-        @DisplayName("비밀번호가 일치하지 않는 경우 예외를 반환한다.")
-        void authorize_user_fail() {
-            //given
-            String email = "a@a.com";
-            String password = "password";
-            given(userRepository.findByEmail(email)).willReturn(Optional.of(new User(email, "original")));
-
-            //when
-            //then
-            assertThatThrownBy(() -> userService.authorizeUser(email, password))
-                    .isInstanceOf(InvalidPasswordException.class);
-        }
-
-        @Test
-        @DisplayName("회원이 존재하지 않을 경우 예외를 반환한다.")
-        void authorize_user_fail_when_user_not_found() {
-            //given
-            String email = "a@a.com";
-            String password = "password";
-            given(userRepository.findByEmail(email)).willReturn(Optional.empty());
-
-            //when
-            //then
-            assertThatThrownBy(() -> userService.authorizeUser(email, password))
-                    .isInstanceOf(UserNotFoundException.class);
-        }
-
     }
 }
